Basic Types/Methods/Clauses, simple recursive visiting and clause matching The type checker is now capable of: - inferring the types of all basic Literals (List and Map are a bit awkward still) - assign types to local variables - create new types (from `deftype`) - create methods on types - add multiple clauses to a method - match a specific clause based on the types of the arguments in a Call
This is major progress, and it shouldn't be long for most of the syntax to be inferrable. However, there are plenty of cases that still need to be covered, including: dynamic receivers, static vs instance methods, inheritance, anything with functors, etc.

The example in `typecheck.cr` shows the inference of `2 + 3.0` as a `Float`, using the type information supplied by the `Integer` type definition above it.