Support typing of loop expressions. Loop expressions are impressively nuanced in how they affect the type environment. In many cases (though not common in normal code), it is possible to determine whether the loop will ever execute, or will stop after some set of iterations. For example,
```
x = :hi
while false
  x = nil
end
```

would never execute the `while` loop body, since the conditional expression is `false`. As such, we know that `x = nil` will never execute, and `Nil` should _not_ be added to the type of `x`.

Even if the evaluation is not known _ahead_ of time, it can sometimes be determined that after an iteration the condition will never succeed. For example,

```
x = 1
y = false
while x
  y = x
  x = nil
end
```

would execute the loop body once (when `x = 1`), but then on the second iteration, `x = nil`, so the loop is guaranteed to _not_ execute. If we just kept typing the loop body without this knowledge, `y` would end up with the type `Integer | Nil`, since even though it was originally assigned `Boolean`, the loop _will_ run at least once, forcing it to be overridden with `Integer`, and then again, when `x = nil`, it would be unioned with `Nil`.

However, with the knowledge that the body will never execute after `x = nil`, we can stop typing the loop body and leave the value of `y` as just `Integer`. This _should_ work for any arbitrary loop body, where the inferrer itself will loop until the environment does not change during an iteration.

Supporting this nuanced typing took a fair amount of effort and could almost certainly be refactored to be more maintainable.