Add basic support for functor definitions. `Functor` is a subclass of `Type`, allowing it to be placed directly into the `scope` of a container, and be compared against other functors/types easily. With that, every functor is a _unique_ type, meaning a functor will only ever be equal to itself.
Note that functors are not immediately typed. Instead, they are just stored in the container's scope, and the typing occurs when the functor gets invoked. At that point, the appropriate clause(s) of the Functor would get matched, have their types inferred using the provided arguments (creating a temporary environment, performing inference, and taking the return type of that evaluation), and unioned to determine the final return type.