Make primitive types per-environment, handle Call receivers. While implementing support for Calls with receivers, I noticed that the primitive types were being shared across tests (i.e., adding a method to a type would persist into other tests). This is obviously not desirable for encapsulation and modularity, so the primitive types have been changed from constants into properties of the Environment type, which is already getting passed around everywhere it is needed, so no new abstractions are needed.
For Call receivers, the checker just types the receiver, then pushes that type onto the `self_stack` and iterates the clauses as before. For receivers that are unions (e.g. `(false || 1)...`), the checker does the same thing individually for each type in the union, then joins them together to determine the final type of the Call.

If any type of the receiver does not have a clause matching the given Call, the checker fails and tells the user about the failure.